Light클래스를 Scene에서 분리
ㅇ
