+++ Client/Debug/LabProject09.log
@@ -1,16 +1,8 @@
-﻿빌드 시작: 2016-05-04 오전 5:09:56
+﻿빌드 시작: 2016-05-04 오전 6:39:24
      1>2 노드의 "D:\ArenaFight\Client\Client\LabProject09.vcxproj" 프로젝트(Build 대상)입니다.
-     1>FxCompile:
-         C:\Program Files (x86)\Microsoft DirectX SDK (June 2010)\Utilities\bin\x86\fxc.exe /nologo /E"main" /T fx_4_0 /Fo "D:\ArenaFight\Client\Client\Debug\Effect.cso" /Od /Zi Effect.fx 
-     1>D:\ArenaFight\Client\Client\Effect.fx(96,21): warning X3206: 'mul': implicit truncation of vector type
-         
-         compilation succeeded; see D:\ArenaFight\Client\Client\Debug\Effect.cso
-       ClCompile:
-         C:\Program Files (x86)\Microsoft Visual Studio 12.0\VC\bin\CL.exe /c /ZI /nologo /W3 /WX- /sdl /Od /Oy- /D WIN32 /D _DEBUG /D _WINDOWS /D _UNICODE /D UNICODE /Gm /EHsc /RTC1 /MDd /GS /fp:precise /Zc:wchar_t /Zc:forScope /Yu"stdafx.h" /Fp"Debug\GameClient.pch" /Fo"Debug\\" /Fd"Debug\vc120.pdb" /Gd /TP /analyze- /errorReport:prompt GameClient.cpp GameFramework.cpp
-         GameFramework.cpp
-         코드를 생성하고 있습니다...
-         건너뛰고 있습니다... (관련된 변경 내용 없음)
-         GameClient.cpp
+     1>ClCompile:
+         C:\Program Files (x86)\Microsoft Visual Studio 12.0\VC\bin\CL.exe /c /ZI /nologo /W3 /WX- /sdl /Od /Oy- /D WIN32 /D _DEBUG /D _WINDOWS /D _UNICODE /D UNICODE /Gm /EHsc /RTC1 /MDd /GS /fp:precise /Zc:wchar_t /Zc:forScope /Yu"stdafx.h" /Fp"Debug\GameClient.pch" /Fo"Debug\\" /Fd"Debug\vc120.pdb" /Gd /TP /analyze- /errorReport:prompt Scene.cpp
+         Scene.cpp
        Link:
          C:\Program Files (x86)\Microsoft Visual Studio 12.0\VC\bin\link.exe /ERRORREPORT:PROMPT /OUT:"D:\ArenaFight\Client\Client\Debug\GameClient.exe" /INCREMENTAL /NOLOGO d3d11.lib d3dx11d.lib d3dcompiler.lib d3dx9d.lib dxerr.lib dxguid.lib dinput8.lib winmm.lib ws2_32.lib kernel32.lib user32.lib gdi32.lib winspool.lib comdlg32.lib advapi32.lib shell32.lib ole32.lib oleaut32.lib uuid.lib odbc32.lib odbccp32.lib /MANIFEST /MANIFESTUAC:"level='asInvoker' uiAccess='false'" /manifest:embed /DEBUG /PDB:"D:\ArenaFight\Client\Client\Debug\GameClient.pdb" /SUBSYSTEM:WINDOWS /TLBID:1 /DYNAMICBASE /NXCOMPAT /IMPLIB:"D:\ArenaFight\Client\Client\Debug\GameClient.lib" /MACHINE:X86 Debug\GameClient.res
          Debug\Camera.obj
@@ -18,6 +10,7 @@
          Debug\Fog.obj
          Debug\GameFramework.obj
          Debug\GameClient.obj
+         Debug\Light.obj
          Debug\Mesh.obj
          Debug\Network.obj
          Debug\Object.obj
@@ -34,4 +27,4 @@
 
 빌드했습니다.
 
-경과 시간: 00:00:00.61
+경과 시간: 00:00:00.42
