+++ Client/Effect.fx
@@ -128,9 +128,9 @@ PS_INPUT VS(VS_INPUT input)
 //													  //
 ////////////////////////////////////////////////////////
 
-#define MAX_LIGHTS		4 
-#define POINT_LIGHT		1.0f
-#define SPOT_LIGHT		2.0f
+#define MAX_LIGHTS			4
+#define POINT_LIGHT			1.0f
+#define SPOT_LIGHT			2.0f
 #define DIRECTIONAL_LIGHT	3.0f
 
 #define _WITH_LOCAL_VIEWER_HIGHLIGHTING
