+++ Client/LabProject09.vcxproj
@@ -93,6 +93,7 @@
     <ClInclude Include="Fog.h" />
     <ClInclude Include="GameFramework.h" />
     <ClInclude Include="GameClient.h" />
+    <ClInclude Include="Light.h" />
     <ClInclude Include="Mesh.h" />
     <ClInclude Include="Object.h" />
     <ClInclude Include="ObjectManager.h" />
@@ -113,6 +114,7 @@
     <ClCompile Include="Fog.cpp" />
     <ClCompile Include="GameFramework.cpp" />
     <ClCompile Include="GameClient.cpp" />
+    <ClCompile Include="Light.cpp" />
     <ClCompile Include="Mesh.cpp" />
     <ClCompile Include="Network.cpp" />
     <ClCompile Include="Object.cpp" />
