+++ Client/Scene.cpp
@@ -4,8 +4,7 @@
 
 CScene::CScene()
 {
-	m_pLights = nullptr;
-	m_pd3dcbLights = nullptr;
+	m_pLight = new CLight();
 }
 
 CScene::~CScene()
@@ -20,7 +19,7 @@ void CScene::BuildObjects(ID3D11Device *pd3dDevice)
 		m_vShaders.push_back(pResourceManager->GetShaderByShaderType((CResourceManager::eShaderType)i));
 	}
 
-	BuildLights(pd3dDevice);
+	m_pLight->BuildLights(pd3dDevice);
 }
 
 void CScene::ReleaseObjects()
@@ -30,17 +29,17 @@ void CScene::ReleaseObjects()
 
 bool CScene::OnProcessingMouseMessage(HWND hWnd, UINT nMessageID, WPARAM wParam, LPARAM lParam)
 {
-	return(false);
+	return false;
 }
 
 bool CScene::OnProcessingKeyboardMessage(HWND hWnd, UINT nMessageID, WPARAM wParam, LPARAM lParam)
 {
-	return(false);
+	return false;
 }
 
 bool CScene::ProcessInput()
 {
-	return(false);
+	return false;
 }
 
 void CScene::AnimateObjects(int StateCnt, ID3D11DeviceContext*pd3dDeviceContext, float fTimeElapsed)
@@ -53,7 +52,7 @@ void CScene::AnimateObjects(int StateCnt, ID3D11DeviceContext*pd3dDeviceContext,
 
 void CScene::Render(ID3D11DeviceContext*pd3dDeviceContext)
 {
-	UpdateLights(pd3dDeviceContext);
+	m_pLight->UpdateLights(pd3dDeviceContext);
 
 	for (auto shader : m_vShaders)
 	{
@@ -63,76 +62,14 @@ void CScene::Render(ID3D11DeviceContext*pd3dDeviceContext)
 
 void CScene::AnimateObjectsAndRender(ID3D11DeviceContext *pd3dDeviceContext, float time)
 {
-	UpdateLights(pd3dDeviceContext);
+	m_pLight->UpdateLights(pd3dDeviceContext);
+
 	for (auto shader : m_vShaders)
 	{
 		shader->AnimateObjectAndRender(pd3dDeviceContext, time);
 	}
 }
 
-void CScene::BuildLights(ID3D11Device *pd3dDevice)
-{
-	m_pLights = new LIGHTS;
-	::ZeroMemory(m_pLights, sizeof(LIGHTS));
-	//게임 월드 전체를 비추는 주변조명을 설정한다.
-	m_pLights->m_d3dxcGlobalAmbient = D3DXCOLOR(0.1f, 0.1f, 0.1f, 1.0f);
 
-	//3개의 조명(점 광원, 스팟 광원, 방향성 광원)을 설정한다.
-	m_pLights->m_pLights[0].m_bEnable = 1.0f;
-	m_pLights->m_pLights[0].m_nType = POINT_LIGHT;
-	m_pLights->m_pLights[0].m_fRange = 40.0f;
-	m_pLights->m_pLights[0].m_d3dxcAmbient = D3DXCOLOR(0.0f, 0.0f, 0.0f, 1.0f);
-	m_pLights->m_pLights[0].m_d3dxcDiffuse = D3DXCOLOR(0.3f, 0.3f, 0.0f, 1.0f);
-	m_pLights->m_pLights[0].m_d3dxcSpecular = D3DXCOLOR(0.1f, 0.1f, 0.1f, 0.0f);
-	m_pLights->m_pLights[0].m_d3dxvPosition = D3DXVECTOR3(0.0f, 100.0f, 0.0f);
-	m_pLights->m_pLights[0].m_d3dxvDirection = D3DXVECTOR3(0.0f, 0.0f, 0.0f);
-	m_pLights->m_pLights[0].m_d3dxvAttenuation = D3DXVECTOR3(1.0f, 0.001f, 0.0001f);
 
-	m_pLights->m_pLights[1].m_bEnable = 1.0f;
-	m_pLights->m_pLights[1].m_nType = SPOT_LIGHT;
-	m_pLights->m_pLights[1].m_fRange = 100.0f;
-	m_pLights->m_pLights[1].m_d3dxcAmbient = D3DXCOLOR(0.1f, 0.1f, 0.1f, 1.0f);
-	m_pLights->m_pLights[1].m_d3dxcDiffuse = D3DXCOLOR(0.3f, 0.3f, 0.0f, 1.0f);
-	m_pLights->m_pLights[1].m_d3dxcSpecular = D3DXCOLOR(0.1f, 0.1f, 0.1f, 0.0f);
-	m_pLights->m_pLights[1].m_d3dxvPosition = D3DXVECTOR3(0.0f, 0.0f, 0.0f);
-	m_pLights->m_pLights[1].m_d3dxvDirection = D3DXVECTOR3(0.0f, 0.0f, 1.0f);
-	m_pLights->m_pLights[1].m_d3dxvAttenuation = D3DXVECTOR3(1.0f, 0.01f, 0.0001f);
-	m_pLights->m_pLights[1].m_fFalloff = 8.0f;
-	m_pLights->m_pLights[1].m_fPhi = (float)cos(D3DXToRadian(40.0f));
-	m_pLights->m_pLights[1].m_fTheta = (float)cos(D3DXToRadian(20.0f));
-
-	m_pLights->m_pLights[2].m_bEnable = 1.0f;
-	m_pLights->m_pLights[2].m_nType = DIRECTIONAL_LIGHT;
-	m_pLights->m_pLights[2].m_d3dxcAmbient = D3DXCOLOR(0.8f, 0.8f, 0.8f, 1.0f);
-	m_pLights->m_pLights[2].m_d3dxcDiffuse = D3DXCOLOR(0.5f, 0.5f, 0.5f, 1.0f);
-	m_pLights->m_pLights[2].m_d3dxcSpecular = D3DXCOLOR(0.0f, 0.0f, 0.0f, 0.0f);
-	m_pLights->m_pLights[2].m_d3dxvDirection = D3DXVECTOR3(0.0f, -1.0f, 0.0f);
-
-	D3D11_BUFFER_DESC d3dBufferDesc;
-	ZeroMemory(&d3dBufferDesc, sizeof(d3dBufferDesc));
-	d3dBufferDesc.Usage = D3D11_USAGE_DYNAMIC;
-	d3dBufferDesc.ByteWidth = sizeof(LIGHTS);
-	d3dBufferDesc.BindFlags = D3D11_BIND_CONSTANT_BUFFER;
-	d3dBufferDesc.CPUAccessFlags = D3D11_CPU_ACCESS_WRITE;
-	D3D11_SUBRESOURCE_DATA d3dBufferData;
-	ZeroMemory(&d3dBufferData, sizeof(D3D11_SUBRESOURCE_DATA));
-	d3dBufferData.pSysMem = m_pLights;
-	pd3dDevice->CreateBuffer(&d3dBufferDesc, &d3dBufferData, &m_pd3dcbLights);
-}
-
-void CScene::ReleaseLights()
-{
-	if (m_pLights) delete m_pLights;
-	if (m_pd3dcbLights) m_pd3dcbLights->Release();
-}
-
-void CScene::UpdateLights(ID3D11DeviceContext *pd3dDeviceContext)
-{
-	D3D11_MAPPED_SUBRESOURCE d3dMappedResource;
-	pd3dDeviceContext->Map(m_pd3dcbLights, 0, D3D11_MAP_WRITE_DISCARD, 0, &d3dMappedResource);
-	LIGHTS *pcbLight = (LIGHTS*)d3dMappedResource.pData;
-	memcpy(pcbLight, m_pLights, sizeof(LIGHTS));
-	pd3dDeviceContext->Unmap(m_pd3dcbLights, 0);
-	pd3dDeviceContext->PSSetConstantBuffers(PS_SLOT_LIGHT, 1, &m_pd3dcbLights);
-}
