+++ Client/Scene.h
@@ -5,40 +5,9 @@
 
 #include "Object.h"
 #include "Shader.h"
-
-#define MAX_LIGHTS		3
-#define POINT_LIGHT		1.0f
-#define SPOT_LIGHT		2.0f
-#define DIRECTIONAL_LIGHT	3.0f
+#include "Light.h"
 
 
-
-
-
-struct LIGHT
-{
-	D3DXCOLOR m_d3dxcAmbient;
-	D3DXCOLOR m_d3dxcDiffuse;
-	D3DXCOLOR m_d3dxcSpecular;
-	D3DXVECTOR3 m_d3dxvPosition;
-	float m_fRange;
-	D3DXVECTOR3 m_d3dxvDirection;
-	float m_nType;
-	D3DXVECTOR3 m_d3dxvAttenuation;
-	float m_fFalloff;
-	float m_fTheta; //cos(m_fTheta)
-	float m_fPhi; //cos(m_fPhi)
-	float m_bEnable;
-	float padding;
-};
-
-struct LIGHTS
-{
-	//상수 버퍼는 크기가 16의 배수가 되어야 한다. 
-	LIGHT m_pLights[MAX_LIGHTS];
-	D3DXCOLOR m_d3dxcGlobalAmbient;
-	D3DXVECTOR4 m_d3dxvCameraPosition;
-};
 
 
 class CScene
@@ -58,17 +27,12 @@ public:
 	void Render(ID3D11DeviceContext*pd3dDeviceContext);			
 
 	// 애니메이션 + 렌더 합치기
-	void AnimateObjectsAndRender(ID3D11DeviceContext* pd3dDeviceContext, float fTimeElapsed);		//추가
-
-	LIGHTS *m_pLights;
-	ID3D11Buffer *m_pd3dcbLights;
-
-	void BuildLights(ID3D11Device *pd3dDevice);
-	void UpdateLights(ID3D11DeviceContext *pd3dDeviceContext);
-	void ReleaseLights();
+	void AnimateObjectsAndRender(ID3D11DeviceContext* pd3dDeviceContext, float fTimeElapsed);	//추가
 
 private:
 	std::vector<CShader*> m_vShaders;
+
+	CLight *m_pLight;
 };
